Show full details when clicking on request info
